Rename graphql to api and add comments
