+++ packages/react-urql/examples/1-getting-started/src/api/graphql.ts
